WatchList - Refactoring.
